Added in my Internet Alarm and the MP3 used for it
Internet alarm does a quick ping to google every second (to avoid DOS-ing the website) to check if I have internet access yet. I have periodically randomly lost access many times in the last few months with no idea of when I will get it back so having a ping that tells me with a fun sound that I have internet is nice!